product by cat changes final

+++ DesiClothing4u/DesiClothing4u.Common/Models/productExt.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -7,10 +8,11 @@ namespace DesiClothing4u.Common.Models
 {
     public class productExt
     {
+        
         public int Id { get; set; }
         public string Name { get; set; }
         public decimal Price { get; set; }
-        public decimal Discount { get; set; }
+        public int Discount { get; set; }
         public string SeoFilename { get; set; }
 
         public string VirtualPath { get; set; }
